Fixed Decoder Issue
Issue was caused by incorrect retrieval of "x" and "context" in Transformer.py's call method. Decoder was passed the context when it should be passed the query sequence.

I also fixed a few other issues that I noticed (i.e. x is not updated in positional encoding but replaced and the square root was not performed on the correct parameter).